LayoutTests/accessibility/ios-simulator/url-test.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
var successfullyParsed = false;
</script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<input id="element1" type="image" src="url-test1.html"><BR>
<a id="element2" href="#url-test2.html">test</a><BR>
<img id="element3" src="url-test3.png" width=100 height=100 alt="test"><BR>


<p id="description"></p>
<div id="console"></div>

<script>

    description("This test ensures that we can retrieve URLs for appropriate elements.");

    if (window.accessibilityController) {

        var element = accessibilityController.accessibleElementById("element1");
        shouldBe("element.url.indexOf('url-test1.html') != -1", "true");

        element = accessibilityController.accessibleElementById("element2");
        shouldBe("element.url.indexOf('url-test2.html') != -1", "true");

        element = accessibilityController.accessibleElementById("element3");
        shouldBe("element.url.indexOf('url-test3.png') != -1", "true");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
